feat: last cloud support

+++ src/main/java/ru/testit/client/model/TestPlanTestPointsWorkItemSearchApiResult.java
@@ -81,7 +81,7 @@ public class TestPlanTestPointsWorkItemSearchApiResult {
 
   public static final String JSON_PROPERTY_DURATION = "duration";
   @jakarta.annotation.Nonnull
-  private Integer duration;
+  private Long duration;
 
   public static final String JSON_PROPERTY_STATE = "state";
   @jakarta.annotation.Nonnull
@@ -230,7 +230,7 @@ public class TestPlanTestPointsWorkItemSearchApiResult {
   }
 
 
-  public TestPlanTestPointsWorkItemSearchApiResult duration(@jakarta.annotation.Nonnull Integer duration) {
+  public TestPlanTestPointsWorkItemSearchApiResult duration(@jakarta.annotation.Nonnull Long duration) {
     this.duration = duration;
     return this;
   }
@@ -243,14 +243,14 @@ public class TestPlanTestPointsWorkItemSearchApiResult {
   @JsonProperty(JSON_PROPERTY_DURATION)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
 
-  public Integer getDuration() {
+  public Long getDuration() {
     return duration;
   }
 
 
   @JsonProperty(JSON_PROPERTY_DURATION)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public void setDuration(@jakarta.annotation.Nonnull Integer duration) {
+  public void setDuration(@jakarta.annotation.Nonnull Long duration) {
     this.duration = duration;
   }
